+++ fbchat/_util.py
